requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
Babel==2.12.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
boto3==1.26.109
botocore==1.29.109
bottle==0.12.25
Brotli==1.0.9
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
common==0.1.2
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
data==0.4
decorator==4.4.2
dual==0.1.0
dynamo3==0.4.10
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.95.0
ffmpy==0.3.0
filelock==3.11.0
fire==0.5.0
Flask==2.2.3
flask-babel==3.1.0
flatbuffers==23.5.26
flywheel==0.5.4
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
funcsigs==1.0.2
future==0.18.3
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.17.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-speech==2.19.0
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
gradio==3.24.1
gradio_client==0.0.8
greenlet==2.0.2
grpcio==1.53.0
grpcio-status==1.53.0
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
imgaug==0.4.0
intel_tensorflow==2.9.1
itsdangerous==2.1.2
jax==0.4.7
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
linkify-it-py==2.0.0
lmdb==1.4.1
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.0.4
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
onnx==1.13.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.9
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==2.0.2
pdf2docx==0.5.6
peewee==3.16.0
Pillow==9.4.0
premailer==3.10.0
proglog==0.1.10
proto-plus==1.22.2
protobuf==3.20.3
prox==0.0.17
pya3==1.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydub==0.25.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
python-docx==0.8.11
python-geoip-python3==1.3
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.2
PyWavelets==1.4.1
pywin32==305.1
PyYAML==6.0
rapidfuzz==2.15.0
rarfile==4.0
rel==0.4.9
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
shapely==2.0.1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
sympy==1.11.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
tight==0.1.0
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
tritonclient==2.32.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
ultralytics==8.0.117
urllib3==1.26.15
uvicorn==0.21.1
visualdl==2.5.1
websocket-client==1.5.3
websockets==11.0.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
x2paddle==1.4.1
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
